Fix non-FreeNAS providers to enable the 'Add' button.
- When using the 'allowBlank' property of extjs, it provided the
functionality to the FreeNAS provider but disabled the 'Add' button of
other providers when all necessary data was entered. This was because
three fields in the FreeNAS provider was 'allowBlank' equal to 'false'
causing the ‘Add’ button to not be ‘enabled’. Upon further development
the 'allowBlank' property is not 'bindable' and was causing issues with
the modal window. No other similar code was found in the
pvemanagerlib.js to replicate this and it was mentioned in a post that
the ‘allowBlank’ property to be bindable was a feature request to come.
This fix allows for all functionality of the providers to work correctly
in the modal window. This fixes issue #36.